Added function to get all tracked shares
also rearranged some function locations into a HoldingController
